fixed database code location
